loopi-restoration highbitdepth fixes

+++ av1/common/restoration.c
@@ -658,7 +658,7 @@ static void loop_sgrproj_filter_tile(uint8_t *data, int tile_idx, int width,
                                rst->rsi->sgrproj_info[tile_idx].xqd, tmpbuf);
   for (i = 0; i < (v_end - v_start); ++i) {
     for (j = 0; j < (h_end - h_start); ++j) {
-      data_p[i * stride + j] = clip_pixel(dat[i * (h_end - h_start) + j]);
+      data_p[i * stride + j] = clip_pixel((int)dat[i * (h_end - h_start) + j]);
     }
   }
 }
@@ -885,7 +885,7 @@ static void loop_sgrproj_filter_tile_highbd(uint16_t *data, int tile_idx,
   for (i = 0; i < (v_end - v_start); ++i) {
     for (j = 0; j < (h_end - h_start); ++j) {
       data_p[i * stride + j] =
-          clip_pixel_highbd(dat[i * (h_end - h_start) + j], bit_depth);
+          clip_pixel_highbd((int)dat[i * (h_end - h_start) + j], bit_depth);
     }
   }
 }
